+++ src/router/email.js
@@ -1,32 +1,12 @@
 const express = require("express");
 const router = express.Router();
-const VerificationCodeService = require("../service/verificationCodeService");
+const emailController = require("../controller/emailController");
 
-router.post("/signup/request", (req, res) => {
-	const email = req.body.email;
+router.post("/signup/request", emailController.postEmailSignupRequest);
 
-	if (VerificationCodeService.sendVerificationCodeByEmail(email))
-		res.sendStatus(200);
-	else res.sendStatus(500);
-});
-
-router.post("/signup/verification", async (req, res) => {
-	const { email, user_code } = req.body;
-
-	try {
-		const result = await VerificationCodeService.verifyUserCode(
-			email,
-			user_code
-		);
-
-		if (result) {
-			req.session.isEmailVerified = true;
-			res.sendStatus(200);
-		} else res.sendStatus(401);
-	} catch (err) {
-		console.error(err);
-		res.sendStatus(500);
-	}
-});
+router.post(
+	"/signup/verification",
+	emailController.postEmailSignupVerification
+);
 
 module.exports = router;
